docker/worker/campusidpScripts/requirements.txt
Jinja2==2.10
pika==1.0.0
PyYAML==5.1
requests==2.21.0
